requirements.txt
asgiref==3.7.2
astroid==3.0.1
asttokens==2.4.1
backcall==0.2.0
decorator==5.1.1
dill==0.3.7
Django==4.2.6
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==2.0.0
ipython==8.16.1
isort==5.12.0
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tomlkit==0.12.1
traitlets==5.12.0
typing_extensions==4.8.0
wcwidth==0.2.8
